Limit width and height for smoel uploads
